+++ planb/transport_exec/models.py
@@ -27,7 +27,7 @@ class Config(models.Model):
         db_table = TABLE_PREFIX  # or '{}_config'.format(TABLE_PREFIX)
 
     def __str__(self):
-        return '{}: exec transport'.format(self.fileset)
+        return 'exec transport {}'.format(self.transport_command)
 
     def clone(self, **override):
         # See: https://github.com/django/django/commit/a97ecfdea8